refactor: QueryString의 특정 키를 다루는 유틸 함수를 모두 분리
